requirements.txt
asgiref==3.6.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-bootstrap-modal-forms==2.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-widget-tweaks==1.4.12
filelock==3.6.0
Flask==2.0.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install==1.3.5
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
ndg-httpsclient==0.5.1
oauthlib==3.2.2
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.6.0
pyOpenSSL==22.0.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
tldextract==3.2.1
typing_extensions==4.2.0
tzdata==2022.7
urllib3==1.26.9
w3lib==1.22.0
webdriver-manager==3.5.4
Werkzeug==2.0.3
zope.interface==5.4.0
